[AS7-1538] Empty OSGi Config Admin property values in AS7 configuration fails
